get_filings/pyproject.toml
[tool.poetry]
name = "etl-13f_get-filings"
version = "0.1.0"
description = "Process to extract 13Fs and load into a DB"
authors = ["Ken Farr <ken@farr.ai>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
edgar3 = "^1.1"
backoff = "^1.10.0"
requests = "^2.22.0"
google-cloud-storage = "^1.31.2"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.7.9"
mypy = "^0.761"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
